+++ src/bindings/module.cpp
@@ -13,6 +13,7 @@ void init_audio_decoder(nb::module_& m);
 void init_video_encoder(nb::module_& m);
 void init_audio_encoder(nb::module_& m);
 void init_video_codec_capabilities(nb::module_& m);
+void init_image_decoder(nb::module_& m);
 
 NB_MODULE(_webcodecs_py, m) {
   m.doc() = "Python bindings for media codecs with WebCodecs-like API";
@@ -28,4 +29,5 @@ NB_MODULE(_webcodecs_py, m) {
   init_video_encoder(m);
   init_audio_encoder(m);
   init_video_codec_capabilities(m);
+  init_image_decoder(m);
 }